setup-requirements.txt
# Minimal requirements for setup script
PyYAML>=6.0
